--- src/components/organisms/MainFooter.stories.tsx
@@ -1,14 +1,0 @@
-import { Story, Meta } from '@storybook/react'
-
-import MainFooter from './MainFooter'
-
-export default {
-  title: 'Organisms/MainFooter',
-  component: MainFooter,
-  argTypes: {}
-} as Meta
-
-const Template: Story = (args) => <MainFooter {...args} />
-
-export const Primary = Template.bind({})
-Primary.args = {}
