+++ src/content/MyCommunitiesContent.tsx
@@ -1,5 +1,5 @@
 import Loading from '@src/components/atoms/Loading'
-import CommunitiesGrid from '@src/components/organisms/CommunitiesGrid'
+import CommunitiesGrid from '@src/fragments/CommunitiesGrid'
 import { useMyCommunities } from '@src/model/Community'
 
 export default function MyCommunitiesContent() {
