+++ src/fragments/CommunityDetails.tsx
