+++ src/fragments/EventDetails.tsx
