+++ src/fragments/MainBar.tsx
@@ -3,7 +3,7 @@ import Navbar from '@src/components/atoms/Navbar'
 import NavbarBrand from '@src/components/atoms/NavbarBrand'
 import NavbarItem from '@src/components/atoms/NavbarItem'
 import Logo from '@src/components/molecules/Logo'
-import UserMenu from '@src/components/organisms/UserMenu'
+import UserMenu from '@src/fragments/UserMenu'
 
 export default function MainBar() {
   return (
